src/main/java/com/mo16/demo4springmvcrest/services/CategoryServiceImpl.java
package com.mo16.demo4springmvcrest.services;

import com.mo16.demo4springmvcrest.api.v1.maodel.CategoryDTO;
import com.mo16.demo4springmvcrest.api.v1.mapppers.CategoryMapper;
import com.mo16.demo4springmvcrest.repositories.CategoryRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.stream.Collectors;

@Service
public class CategoryServiceImpl implements CategoryService {

    private final CategoryRepository categoryRepository;
    private final CategoryMapper mapper;

    @Autowired
    public CategoryServiceImpl(CategoryRepository categoryRepository, CategoryMapper mapper) {
        this.categoryRepository = categoryRepository;
        this.mapper = mapper;
    }

    @Override
    public List<CategoryDTO> getAllCategories() {
        return categoryRepository
                .findAll()
                .stream()
                .map(mapper::categoryDtoFromCategory)
                .collect(Collectors.toList())
                ;
    }

    @Override
    public CategoryDTO getCategoryByName(String name) {
        return mapper.categoryDtoFromCategory(categoryRepository.findByName(name));
    }
}
